refactor: reorganize dlt code (#1367)
