+++ catalog/py_package/catalog_build/load/__init__.py
@@ -1,0 +1,1 @@
+from .load import *
