+++ catalog/py_package/catalog_build/load/catalog_source.py
@@ -1,0 +1,56 @@
+from pathlib import Path
+
+import dlt
+import pandas as pd
+
+from ..utils import get_db_path_string
+
+
+@dlt.resource(name="outbreak_taxa", write_disposition="replace")
+def outbreak_taxa(df: pd.DataFrame):
+    yield df
+
+
+@dlt.resource(name="organism_taxa", write_disposition="replace")
+def organism_taxa(df: pd.DataFrame):
+    yield df
+
+
+@dlt.resource(name="assembly_taxa", write_disposition="replace")
+def assembly_taxa(df: pd.DataFrame):
+    yield df
+
+
+@dlt.source
+def catalog_taxa(
+    *,
+    assembly_taxa_df: pd.DataFrame,
+    organism_taxa_df: pd.DataFrame,
+    outbreak_taxa_df: pd.DataFrame,
+):
+    return [
+        assembly_taxa(assembly_taxa_df),
+        organism_taxa(organism_taxa_df),
+        outbreak_taxa(outbreak_taxa_df),
+    ]
+
+
+def load_catalog_taxa(
+    temp_folder_path: Path,
+    *,
+    assembly_taxa_df: pd.DataFrame,
+    organism_taxa_df: pd.DataFrame,
+    outbreak_taxa_df: pd.DataFrame,
+):
+    pipeline = dlt.pipeline(
+        pipeline_name="catalog_data",
+        destination=dlt.destinations.duckdb(get_db_path_string(temp_folder_path)),
+        dataset_name="catalog_data",
+    )
+    pipeline.run(
+        catalog_taxa(
+            assembly_taxa_df=assembly_taxa_df,
+            organism_taxa_df=organism_taxa_df,
+            outbreak_taxa_df=outbreak_taxa_df,
+        )
+    )
